backend/requirements.txt
Flask
google-generativeai
chromadb
requests
feedparser
python-dotenv
gunicorn
html2text
playwright
beautifulsoup4